refactor(modal): Move reusable useOpenClose hook
Hook is useful for other components and can be
moved to the common hooks.
